requirements.txt
streamlit
python-dotenv
huggingface-hub